pyproject.toml
[project]
name = "langchain-fastapi-starter"
version = "0.1.0"
description = "Minimal FastAPI service with LangChain"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.111.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "langchain>=0.2.0,<0.4.0",
    "langchain-openai>=0.1.0,<0.3.0",
    "pydantic-settings>=2.3.0,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "langchain-ollama>=0.3.10",
]

[dependency-groups]
dev = [
    "ruff>=0.6.9,<1.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
